+++ CHANGELOG.md
@@ -51,6 +51,7 @@ MunicipalityListener, BoardListener and their tests.
 ### Changed
 - [NSEK-138](https://jira.itkdev.dk/browse/NSEK-138): Updating Cypress to 6.5.0 so the same version is used project wide.
 - [TVIST1-144](https://jira.itkdev.dk/browse/TVIST1-144): Switched to `itk-dev/openid-connect-bundle` for AD and CLI login.
+- [TVIST1-670](https://jira.itkdev.dk/browse/TVIST1-670): Updated `${agenda.items}` template macro.
 
 ### Fixed
 - [NSEK-104](https://jira.itkdev.dk/browse/NSEK-104): Added missing database variables to .env file.